FinanceTracker/backend/requirements.txt
asgiref>=3.7.0
Django==4.2.30
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.1
PyJWT==2.8.0
sqlparse==0.4.4
tzdata==2024.1
django-filter
gunicorn
psycopg2-binary
dj-database-url
whitenoise
